refactor: 헥사고날 구조에 맞도록 패키지 리팩토링 (#12)
* refactor: 패키지 리팩토링

* refactor: spotless 적용

---------
