+++ api/src/main/java/com/flab/ccinside/api/trendingpost/application/PostMapper.java
@@ -1,6 +1,6 @@
 package com.flab.ccinside.api.trendingpost.application;
 
-import com.flab.ccinside.api.trendingpost.application.port.out.PostData;
+import com.flab.ccinside.api.trendingpost.application.port.out.post.PostData;
 import com.flab.ccinside.api.trendingpost.domain.Post;
 import org.springframework.stereotype.Component;
