+++ api/src/main/java/com/flab/ccinside/api/trendingpost/application/TrendingPostMapper.java
@@ -1,6 +1,6 @@
 package com.flab.ccinside.api.trendingpost.application;
 
-import com.flab.ccinside.api.trendingpost.application.port.out.TrendingPostData;
+import com.flab.ccinside.api.trendingpost.application.port.out.trendingpost.TrendingPostData;
 import com.flab.ccinside.api.trendingpost.domain.TrendingPost;
 import java.util.List;
 import org.springframework.stereotype.Component;
