+++ api/src/main/java/com/flab/ccinside/api/trendingpost/application/port/in/AddPostViewCountCommand.java
@@ -2,6 +2,4 @@ package com.flab.ccinside.api.trendingpost.application.port.in;
 
 import com.flab.ccinside.api.trendingpost.domain.Post;
 
-public record AddPostViewCountCommand(Post post) {
-
-}
+public record AddPostViewCountCommand(Post post) {}
