+++ api/src/main/java/com/flab/ccinside/api/trendingpost/application/port/in/TrendingPostUseCase.java
@@ -1,7 +1,7 @@
 package com.flab.ccinside.api.trendingpost.application.port.in;
 
-import com.flab.ccinside.api.trendingpost.application.port.out.TrendingPostData;
-import com.flab.ccinside.api.trendingpost.application.port.out.UnitTime;
+import com.flab.ccinside.api.trendingpost.application.port.out.trendingpost.TrendingPostData;
+import com.flab.ccinside.api.trendingpost.application.port.out.trendingpost.UnitTime;
 import java.util.List;
 
 public interface TrendingPostUseCase {
