+++ api/src/main/resources/application-local.yml
@@ -12,3 +12,4 @@ spring:
     redis:
       host: localhost
       port: 6381
+      database: 1
